add check_possible_fhir_resources
