+++ src/anamnesisai/__init__.py
@@ -2,8 +2,6 @@
 
 from importlib import metadata as importlib_metadata
 
-from anamnesisai.core import extract_fhir_openai
-
 
 def get_version() -> str:
     """Return the program version."""
@@ -18,8 +16,3 @@ version = get_version()
 __version__ = version
 __author__ = "Satarupa Deb, Ivan Ogasawara"
 __email__ = "satarupa2212@gmail.com, ivan.ogasawara@gmail.com"
-
-
-__all__ = [
-    "extract_fhir_openai",
-]
